Sun, Mar 12, 2023, 12:44 AM +03:00

+++ backend/api.py
@@ -158,7 +158,7 @@ def remove():
     if len(query) != 1:
         return jsonify("Something went wrong", 500)
 
-    db.execute("DELETE FROM items WHERE item_id = ?", key)
+    db.execute("DELETE FROM items WHERE user_id = ? AND item_id = ?", key)
 
 
     for row in query:
